+++ src/main/groovy/com/trevorism/gcloud/webapi/model/filtering/ComplexFilter.groovy
@@ -1,0 +1,8 @@
+package com.trevorism.gcloud.webapi.model.filtering
+
+
+class ComplexFilter {
+    String type // simple | and | or   (simple is the default)
+    List<SimpleFilter> simpleFilters = []
+    List<ComplexFilter> complexFilters = []
+}