+++ src/main/groovy/com/trevorism/gcloud/webapi/model/filtering/FilterConstants.groovy
@@ -1,0 +1,21 @@
+package com.trevorism.gcloud.webapi.model.filtering
+
+class FilterConstants {
+    static final String TYPE_STRING = "string"
+    static final String TYPE_NUMBER = "number"
+    static final String TYPE_DATE = "date"
+    static final String TYPE_BOOLEAN = "boolean"
+
+    static final String AND = "and"
+    static final String OR = "or"
+
+    static final String OPERATOR_EQUAL = "="
+    static final String OPERATOR_GREATER_THAN = ">"
+    static final String OPERATOR_LESS_THAN = "<"
+    static final String OPERATOR_GREATER_THAN_OR_EQUAL = ">="
+    static final String OPERATOR_LESS_THAN_OR_EQUAL = "<="
+    static final String OPERATOR_NOT_EQUAL = "!="
+
+
+
+}