+++ src/main/groovy/com/trevorism/gcloud/webapi/model/filtering/SimpleFilter.groovy
@@ -1,0 +1,11 @@
+package com.trevorism.gcloud.webapi.model.filtering
+
+import groovy.transform.ToString
+
+@ToString
+class SimpleFilter {
+    String type // string | number | date   (string is the default)
+    String field
+    String operator
+    String value
+}