+++ src/main/groovy/com/trevorism/gcloud/webapi/model/paging/Page.groovy
@@ -1,0 +1,7 @@
+package com.trevorism.gcloud.webapi.model.paging
+
+class Page {
+    int page
+    int pageSize
+    int limit
+}