+++ src/main/groovy/com/trevorism/gcloud/webapi/model/sorting/ComplexSort.groovy
@@ -1,0 +1,6 @@
+package com.trevorism.gcloud.webapi.model.sorting
+
+class ComplexSort {
+    List<Sort> sorts = []
+
+}