+++ src/main/groovy/com/trevorism/gcloud/webapi/model/sorting/Sort.groovy
@@ -1,0 +1,6 @@
+package com.trevorism.gcloud.webapi.model.sorting
+
+class Sort {
+    String field
+    boolean descending
+}