+++ src/main/groovy/com/trevorism/gcloud/webapi/service/EntityList.groovy
@@ -1,0 +1,17 @@
+package com.trevorism.gcloud.webapi.service
+
+import com.google.cloud.datastore.Entity
+import com.google.cloud.datastore.QueryResults
+
+class EntityList {
+
+    private final QueryResults<Entity> results
+
+    EntityList(QueryResults<Entity> results){
+        this.results = results
+    }
+
+    List<Entity> toList(){
+        results.collect ()
+    }
+}