+++ src/main/groovy/com/trevorism/gcloud/webapi/service/FilterService.groovy
@@ -1,0 +1,9 @@
+package com.trevorism.gcloud.webapi.service
+
+import com.trevorism.gcloud.webapi.model.filtering.ComplexFilter
+
+interface FilterService {
+
+    def filter(ComplexFilter request, String kind)
+
+}