+++ src/main/groovy/com/trevorism/gcloud/webapi/service/PagingService.groovy
@@ -1,0 +1,10 @@
+package com.trevorism.gcloud.webapi.service
+
+
+import com.trevorism.gcloud.webapi.model.paging.Page
+
+interface PagingService {
+
+    def page(Page pagingRequest, String kind)
+
+}