+++ src/main/groovy/com/trevorism/gcloud/webapi/service/SortService.groovy
@@ -1,0 +1,8 @@
+package com.trevorism.gcloud.webapi.service
+
+import com.trevorism.gcloud.webapi.model.sorting.ComplexSort
+
+interface SortService {
+
+    def sort(ComplexSort request, String kind)
+}